+++ src/main/java/roomescape/dto/ReservationRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
